% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fin}
\alias{fin}
\title{Finland}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5464 rows and 573 columns.}
\source{
\url{http://www.oecd.org/skills/piaac/}
}
\usage{
fin
}
\description{
Datasets from public-use files of the OECD Programme for the International
Assesment of Adult competencies.
}
\keyword{datasets}

