% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gbr}
\alias{gbr}
\title{United Kingdom (England and Northern Ireland)}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8892 rows and 562 columns.}
\source{
\url{http://www.oecd.org/skills/piaac/}
}
\usage{
gbr
}
\description{
Datasets from public-use files of the OECD Programme for the International
Assesment of Adult competencies.
}
\keyword{datasets}

