% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPIAAC.R
\docType{package}
\name{RPIAAC}
\alias{RPIAAC}
\alias{RPIAAC-package}
\title{RPIAAC: datasets from the Survey of Adult Skills (PIAAC)}
\description{
This packages provides the dta contained in the public-use files of the first
and second round of the PIAAC.
}

