% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{svn}
\alias{svn}
\title{Slovenia}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5331 rows and 604 columns.}
\source{
\url{http://www.oecd.org/skills/piaac/}
}
\usage{
svn
}
\description{
Datasets from public-use files of the OECD Programme for the International
Assesment of Adult competencies.
}
\keyword{datasets}

