% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{variable_coding}
\alias{variable_coding}
\title{Coding scheme for a PIAAC variable}
\usage{
variable_coding(var)
}
\arguments{
\item{var}{Name of a variable}
}
\value{
A dataframe with the coding scheme
}
\description{
Coding scheme for a PIAAC variable
}
\examples{
variable_coding("AGE_R")
variable_coding("AGE5LFS")
}

