% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{variable_info}
\alias{variable_info}
\title{Return information on PIAAC variables}
\usage{
variable_info(var)
}
\arguments{
\item{var}{Name of a variable}
}
\value{
A list with metadata information
}
\description{
Return information on PIAAC variables
}
\examples{
variable_info("AGE_R")
variable_info("AGE5LFS")
}

